Refactor okLab methods
